fix: checking for updates
The GitHub CLI slightly altered how it checks extensions for upgrades.
The new way will still log that the extension is up-to-date, but will
_always_ log "✓ Would have upgraded extension" when the --dry-run flag
is enabled.

See https://github.com/cli/cli/commit/88cae9f5be67989be191cd4a1fe043ddbeafe629

All we need is to slightly alter the message text we're checking for in
order to get this working again.